pyproject.toml
[project]
name = "transcribe"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.15.0",
    "lingua-language-detector>=2.1.0",
    "openai>=1.78.1",
    "python-dotenv>=1.1.0",
]

[dependency-groups]
dev = [
    "ipdb>=0.13.13",
]
